WIP: generalization SendDocumentStoreBatchTransaction task
